sort the rename-command hash
It's a hash. hashes can't be trusted to always stay in the same order,
and sometimes that order does change, which means the config file
changes, which means redis restarts. Which has been happening to me!

Sorting the hash first means it'll always be in the same order, no
matter what. The order of the rename-command statements don't actually
matter, so just sort them and they'll always be consistent!