handle "false" as the save class parameter
this should make it so you can easily use a hiera override on the 'save'
parameter in a boolean fashion. for instance, I have
a redis::config::save key in my global hiera level, but for certain
things I don't actually want redis to save to disk (my CI machines don't
need it, for instance) so I want to be able to override it. Using an
empty array works, but it's not boolean false, so I can't test the
parameter elsewhere